docagent/requirements.txt
flask>=3.0.0
langgraph>=0.0.20
langchain-core>=0.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
PyPDF2>=3.0.0
markdown>=3.5.0
weasyprint>=58.0
python-dotenv>=1.0.0
anthropic>=0.40.0